backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.10
python-dateutil==2.8.2